Moving Slack specific parts of the model to its own subpackage
