+++ src/main/java/com/offdk/play/model/slack/SubCommand.java
@@ -1,4 +1,4 @@
-package com.offdk.play.model;
+package com.offdk.play.model.slack;
 
 public enum SubCommand {
   Challenge, Accept, Score, Top;
