+++ src/main/java/com/offdk/play/web/SlackCommandController.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
-import com.offdk.play.model.SlackCommand;
+import com.offdk.play.model.slack.SlackCommand;
 
 @Controller
 public class SlackCommandController {
